Refactor metric outer loop tests to separate projects to avoid timeout issues
